+++ andes/__init__.py
@@ -10,8 +10,8 @@ from andes import utils      # NOQA
 from andes import variables  # NOQA
 
 from andes.main import config_logger, run, load, prepare  # NOQA
-from andes.system import System        # NOQA
-from andes.utils.paths import get_case # NOQA
+from andes.system import System                    # NOQA
+from andes.utils.paths import get_case, list_cases # NOQA
 
 
 __author__ = 'Hantao Cui'